+++ svelte/src/components/wrapper.tsx
@@ -1,14 +1,14 @@
 import { API } from '@storybook/api';
 
 import { STORY_CHANGED, STORY_RENDERED } from '@storybook/core-events';
-import React, { Key, useState } from 'react';
+import React, { useState } from 'react';
 
 // Component
 import InteractionPanel from './panel';
 // Helper
 import { DEFINE_ACTORS, WS_PORT } from '../helpers/constants';
 // Type
-import { Actors, Framework } from '../../types';
+import { Actors, Argument, Framework, ObjectExpression } from '../../types';
 
 // WebSocket client
 const ws = new WebSocket(`ws://localhost:${WS_PORT}`);
@@ -17,6 +17,9 @@ export default ({ api, active }: { api: API, active: boolean }) => {
   const [ actors, setActors ] = useState<Actors>();
   const [ renderedState, setRenderedState ] = useState<boolean>(false);
   const [ framework, setFramework ] = useState<Framework>('svelte');
+  const [ componentName, setComponentName ] = useState<string>('');
+  const [ extension, setExtension ] = useState<string>('');
+  const [ props, setProps ] = useState<ObjectExpression['properties']>([]);
   // Listen for future story selections
   api.on(STORY_RENDERED, () => {
     const currentStory = api.getCurrentStoryData();
@@ -33,11 +36,22 @@ export default ({ api, active }: { api: API, active: boolean }) => {
   api.on(STORY_CHANGED, () => {
     setRenderedState(false);
   });
+  ws.onmessage = (e) => {
+    const response = JSON.parse(e.data);
+    setComponentName(response.fileName);
+    setExtension(response.fileExtension);
+    setProps(response.exportDeclarations);
+  };
   return (
     <InteractionPanel
       active={active as boolean}
       actors={actors as Actors}
       storyRendered={renderedState}
-      framework={framework} />
+      framework={framework}
+      component={{
+        componentName,
+        extension,
+        props: props as ObjectExpression['properties'],
+      }} />
   )
 };