+++ svelte/src/helpers/index.ts
@@ -1,4 +1,4 @@
 export * from './functions';
 export * from './constants';
 export * from './eventListeners';
-export * from './jestAstMapper';
+export { default as Jest } from './jestAstMapper';