+++ svelte/src/helpers/preset_fileParser.ts
@@ -3,8 +3,9 @@ import fs from 'fs';
 import { parse as svelteParser } from 'svelte/compiler';
 import { parse as babelParser } from '@babel/parser';
 import traverse from '@babel/traverse';
+import generate from '@babel/generator';
 
-import { Actors, UserEvent } from '../../types';
+import { ObjectExpression } from '../../types';
 import { DEFINE_ACTORS } from './constants';
 
 export const getFileContent = (filePath: string) => {
@@ -15,35 +16,41 @@ export const getFileContent = (filePath: string) => {
 
 type FileExtensionType = 'svelte' | 'stories.svelte';
 
-let results: UserEvent[] = [];
-
 const analyseJsTsScriptBlock = (code: string) => {
   const ast = babelParser(code, {
     sourceType: 'module',
     plugins: ['typescript'],
   });
 
+  const exportedProps: ObjectExpression['properties'] = [];
   // Reference: https://github.com/jamiebuilds/babel-handbook/blob/master/translations/en/plugin-handbook.md#toc-asts
+  // Most of the code for traverse are from ChatGPT as I knew nothing about AST 🥲
   traverse(ast, {
-    CallExpression(path) {
-      if (
-        path.node.callee.type === 'MemberExpression' &&
-        path.node.callee.property.type === 'Identifier' &&
-        path.node.callee.property.name === 'addEventListener'
-      ) {
-        const [ eventType, handler ] = path.node.arguments;
-        if (eventType.type === 'StringLiteral') {
-          results.push({
-            type: 'script',
-            eventName: eventType.value,
-            element: (path.node.callee.object as any)?.name,
-          });
+    ExportNamedDeclaration(path) {
+      const declaration = path.node.declaration;
+      if (declaration?.type === 'VariableDeclaration') {
+        for (const decl of declaration.declarations) {
+          if (
+            decl.id.type === 'Identifier' &&
+            (declaration.kind === 'let' || declaration.kind === 'const')
+          ) {
+            const name = decl.id.name;
+            let defaultValue: string = '';
+
+            if (decl.init) {
+              const { code } = generate(decl.init);
+              defaultValue = code;
+            }
+            exportedProps.push({ key: name, value: defaultValue });
+          }
         }
       }
     },
   });
+
+  return exportedProps;
 };
-const analyseSvelteCode = (code: string, actors: Actors) => {
+const analyseSvelteCode = (code: string) => {
   const parsed = svelteParser(code);
   const scriptBlock = parsed.instance;
   if (!scriptBlock) {
@@ -52,42 +59,7 @@ const analyseSvelteCode = (code: string, actors: Actors) => {
   const scriptCode = code.slice((scriptBlock as any).content.start, (scriptBlock as any).content.end);
 
   // Analyse the script block
-  analyseJsTsScriptBlock(scriptCode);
-
-  // Analyse the template block
-  const templateBlock = parsed.html;
-  const walk = (node: any) => {
-    if (node.type === 'InlineComponent' && node.attributes) {
-      for (const attr of node.attributes) {
-        if (attr.type === 'EventHandler') {
-
-        }
-      }
-    }
-    if ((node.type === 'Element') && node.attributes) {
-      console.log('node', node);
-      for (const attr of node.attributes) {
-        if (attr.type === 'Attribute') {
-          results.push({
-            type: 'template',
-            eventName: attr.name,
-            element: node.name,
-            handler: attr.expression?.name,
-            location: attr.expression?.start
-              ? {
-                  line: attr.expression.start.line,
-                  column: attr.expression.start.column,
-                }
-              : undefined,
-          });
-        }
-      }
-    }
-
-    if (node.children) node.children.forEach(walk);
-  };
-
-  walk(templateBlock);
+  return analyseJsTsScriptBlock(scriptCode);
 };
 
 type AnalyseCodeType = {
@@ -110,8 +82,5 @@ export const analyseCode = ({ filePath, fileContent = null, parameters }: Partia
     process.exit(1);
   }
 
-  // Babel
-  analyseSvelteCode(content, parameters[DEFINE_ACTORS]);
-  console.log('results', results);
-  // Svelte compiler
+  return analyseSvelteCode(content);
 }