+++ svelte/types/index.ts
@@ -5,6 +5,17 @@ export type EventType = 'click' | 'hover' | 'input' | 'change';
 export type Framework = 'svelte' | 'react' | 'angular';
 
 export type JestQuery = 'queryByTestId' | 'queryAllByTestId';
+export type JestLifeCycleFunction =
+  | JestQuery
+  | 'beforeAll'
+  | 'beforeEach'
+  | 'afterEach'
+  | 'afterAll'
+  | 'render'
+  | 'waitFor'
+  | 'waitForElementToBeRemoved'
+  | 'userEvent'
+  | 'fireEvent';
 
 export type EnrichedStory = {
   id: string;
@@ -36,7 +47,16 @@ export type UserEventResult = {
   };
 };
 
-export type JavascriptUntouchedKeyword = 'await' | 'function' | 'new' | 'break' | 'continue' | 'const' | 'let';
+export type JavascriptUntouchedKeyword =
+  | 'await'
+  | 'function'
+  | 'new'
+  | 'break'
+  | 'continue'
+  | 'const'
+  | 'let'
+  | 'import'
+  | 'export';
 export type FunctionExpression = {
   type: 'arrow_function';
   params: Array<any>;
@@ -44,18 +64,46 @@ export type FunctionExpression = {
 };
 export type ObjectExpression = {
   type: 'object';
-  properties: {
-    key: string;
-    value: Argument;
-  };
+  properties:
+    | { key: string; value: Argument }
+    | Array<{ key: string; value: Argument }>
 };
 export type Argument = HTMLElement | Element | string | number | ObjectExpression | FunctionExpression;
-export type JestMemberExpression = {
+export type JestMemberExpression<K extends 'logic' | 'declaration'> = {
   object: string;
-  property: string;
+  property: K extends 'logic' ? string : string | undefined;
   arguments?: Array<Argument>;
 };
+export type ImportExportKind = 'named' | 'default';
+export  type JestDeclarationExpression = {
+  importDeclaration: Array<{
+    declarationSyntax: ImportExportKind;
+    keywords: Set<Omit<JestLifeCycleFunction, JestQuery>>;
+    source: JestImportPath;
+  }>;
+  variableDeclaration: Array<{
+    declarationSyntax: ImportExportKind;
+    keywords: Set<JestQuery>;
+    callee: JestMemberExpression<'declaration'>;
+  }>;
+}
 export type JestExpressionStatement = {
   keyword?: JavascriptUntouchedKeyword;
-  callee: JestMemberExpression;
+  callee: JestMemberExpression<'logic'>;
+};
+
+export type JestImportPath =
+  '@jest/global'
+  | '@testing-library/jest-dom/extend-expect'
+  | '@testing-library/svelte'
+  | '@testing-library/react'
+  | '@testing-library/angular'
+  | '@testing-library/user-event';
+type FromType<T> = T extends 'importDeclaration' ? JestImportPath : 'render';
+export type DeclarationTemplate = {
+  [x in JestLifeCycleFunction]: {
+    kind: 'importDeclaration' | 'variableDeclaration';
+    from: FromType<DeclarationTemplate[x]['kind']>;
+    declarationSyntax: ImportExportKind;
+  };
 };