+++ svelte/CHANGELOG.md
@@ -38,4 +38,6 @@ What we can do is define our own “actors and scenes” from Storybook paramete
 - Dynamic height for the recorded step list when they expand the maximum height of its container element.
 - Debounce input event handler for 500ms whilst keeping other event handlers in tact.
 - Updated the UI to display framework logo and test framework selection for future iteration on other test frameworks.
+## [20/04/2025]
+- Revived the preset side to send over the file base name as well as the results of parsing svelte file content to get the initial values from export statements.
 <br />